Re-factored some of the in-memory context code
